refactor(scripts): split verify-docs.py into verify_docs/ package
812-line script with 8 concerns mixed together (keywords / changelog
xref / screenshot freshness / stamp mutators / required links /
Marketplace sync / asset inventory / shared infra) split into one
module per concern. Entry point stays at scripts/verify-docs.py so the
pre-commit hook and CI invocation are unchanged; it is now a thin shim
(~40 lines) that imports verify_docs.cli.main.

New package layout:
  scripts/verify-docs.py        # thin entry (argparse + sys.exit)
  scripts/verify_docs/
    __init__.py                 # module map docstring
    paths.py                    # REPO_ROOT + file constants
    report.py                   # Finding / Report types
    features.py                 # load + flatten iterator
    plugin_xml.py               # <description> CDATA extractor
    git_utils.py                # rev-parse / merge-base / changed-since / sha256
    keywords.py                 # Invariant 1
    changelog.py                # Invariant 2
    screenshots.py              # Invariant 3 (freshness + content hash)
    stamps.py                   # --update-hashes / --restamp mutators
    required_links.py           # Invariant 4
    marketplace.py              # Invariant 6
    inventory.py                # Invariant 5
    cli.py                      # argparse + orchestration

Behavioral contract preserved:
- scripts/verify-docs.py accepts --update-hashes and --restamp exactly
  as before; pre-commit + CI invocations unchanged.
- Report output format identical (same prefixes, same message bodies).
- Function semantics untouched; regex rewrites and git helpers copied
  verbatim to their new homes.
- pyright clean (verified via `uv run --with pyright pyright`).

Private helpers (underscore-prefixed, e.g. _check_one_screenshot,
_handle_orphan_stamp) stay private to their module. Public helpers in
git_utils got their leading underscore dropped (primary_ref, rev_parse,
changed_since, etc.) since they are now legitimate cross-module entry
points.

Side effect of working on a feature branch: three accent + font
screenshot last_verified_sha stamps re-stamped to current HEAD (their
previous SHAs orphaned when release/v2.5.2 was squash-merged).
